Command/Instruction unification (#3862)
* Begin command and instruction unification within Pulse. Instructions will migrate into the new instructions submodule of pulse. This begins by moving only the Instruction class. The Instruction ``command`` attribute is deprecated in favor of ``operands``, and references to commands are updated.

* Remove deprecation warning from commands/instruction.py because the commands/__init__ can have a deprecation warning after the rest is implemented

* Revert commands/__init__ because changes there are not necessary.

* Fill in first positional argument in instruction. What was command is now just 'duration', and any other operands only exist for an instruction instance

* Continue adding documentation and notes

Fix up hash function

* Fix type hint from List[Channel] to Channel

* Update repr to match new init signature

* Update qiskit/pulse/instructions/__init__.py


* Update init file for the instruction module to be more clear about the operands

* Make Instruction an ABC

* update instruction comment

* Update docs

* Update year in copyright header

* Revert Instruction as ABC because the Instructions which haven't been migrated yet do not support the abstractmethods

* Update __eq__ to support both styles until migration is complete

* Update __hash__ to support both styles until migration is complete

* Fixup style

* Update TODO comment

* Use self.command instead of init arg  if the first arg is a command, for readability

* Make repr unchanged for nonmigrated commands

* Small bugfix

* Update instructions documentation

* Relax requirement for channel operand

* API reference to Instruction in release notes

* Fix releasenote

* Silly reno style
