+++ releasenotes/notes/unify-instructions-and-commands-aaa6d8724b8a29d3.yaml
@@ -1,0 +1,24 @@
+---
+features:
+  - |
+    There has been a significant simplification to the style in which Pulse
+    instructions are built.
+
+    With the previous style, ``Command`` s were called with channels to make
+    an :py:class:`~qiskit.pulse.instruction.Instruction`. The usage of both
+    commands and instructions was a point of confusion. This was the previous
+    style::
+
+        sched += Delay(5)(DriveChannel(0))
+        sched += ShiftPhase(np.pi)(DriveChannel(0))
+
+    or, equivalently (though less used)::
+
+        sched += DelayInstruction(Delay(5), DriveChannel(0))
+        sched += ShiftPhaseInstruction(ShiftPhase(np.pi), DriveChannel(0))
+
+    Now, rather than build a command *and* an instruction, each command has
+    been migrated into an instruction::
+
+        sched += Delay(5, DriveChannel(0))
+        sched += ShiftPhase(np.pi, DriveChannel(0))
